[refactor] 주문관련 전체코드

+++ src/main/java/com/example/whattheeat/dto/OrderResponseDto.java
@@ -3,6 +3,8 @@ package com.example.whattheeat.dto;
 import com.example.whattheeat.enums.OrderStatus;
 import lombok.Getter;
 
+import java.math.BigDecimal;
+
 @Getter
 public class OrderResponseDto {
 
@@ -11,11 +13,11 @@ public class OrderResponseDto {
     private Long menuId;
     private Long shopId;
     private int quantity;
-    private int totalPrice;
+    private BigDecimal totalPrice;
     private String address;
     private OrderStatus orderStatus;
 
-    public OrderResponseDto(Long id, Long customerId, Long menuId, Long shopId, int quantity, int totalPrice, String address, OrderStatus orderStatus) {
+    public OrderResponseDto(Long id, Long customerId, Long menuId, Long shopId, int quantity, BigDecimal totalPrice, String address, OrderStatus orderStatus) {
         this.id = id;
         this.customerId = customerId;
         this.menuId = menuId;
